+++ modules/generatable.rb
@@ -1,4 +1,4 @@
-module Encryption
+module Generatable
   def random_key
     rand(00000..99999).to_s.rjust(5, "0")
   end
