+++ spec/spec_helper.rb
@@ -4,3 +4,6 @@ SimpleCov.start
 require 'pry'
 require 'date'
 require './lib/enigma'
+require './modules/encryptable'
+require './modules/decryptable'
+require './modules/generatable'
